model loading

+++ BundtEngineCPP/BundtLowLevelEngine/bunkan.cpp
@@ -9,7 +9,7 @@
 int start(
 	Vector3* vertices,
 	uint32_t verticesCount,
-	uint16_t* indices,
+	uint32_t* indices,
 	uint32_t indicesCount,
 	Vector3* colors,
 	Vector2* texCoords)
