+++ BundtEngineCPP/BundtLowLevelEngine/bunkan.h
@@ -8,7 +8,7 @@
 extern "C" __declspec(dllexport) int __cdecl start(
 	Vector3* vertices,
 	uint32_t verticesCount,
-	uint16_t* indices,
+	uint32_t* indices,
 	uint32_t indicesCount,
 	Vector3* colors,
 	Vector2* texCoords);
